requirements.txt
django==5.2.6
dadata==25.7.0
requests==2.32.5
python-decouple==3.8
bs4==0.0.2
beautifulsoup4==4.13.5
djangorestframework==3.16.1